refactor: clean up file structure
