+++ frontend/src/components/pages/events/EventDetails.tsx
@@ -3,7 +3,7 @@ import LoginRequired from "@components/authz/LoginRequired";
 import Breadcrumbs from "@components/Breadcrumbs";
 import LabeledIcon from "@components/LabeledIcon";
 import * as components from "@components/markdown/components";
-import PermissionRequired from "@components/authz/PermissionRequired";
+import { PermissionRequired } from "@components/authz";
 import { EVENT_SIGN_OFF, EVENT_SIGN_UP } from "@graphql/events/mutations";
 import { GET_EVENT } from "@graphql/events/queries";
 import { GET_USER } from "@graphql/users/queries";
